Java P2P Start
Alright. In this current version, I have a couple of Jobjects put into
the frame to give you guys a basic understanding of how some of that
works. I'll be updating the UI to hold a list of currently connected
peers to the server.
Right now, I'm worried that we will only be able to make one peer
connect, as the while loop doesn't read off the same system.out comment
more than once. There is also the issue with testing. Similar to the
chatroom example package I included, testing this is as simple as
starting the server and then the client/peer. The problem is how do we
run multiple peers on the localhost for testing. I don't think it will
be an issue, Eclipse should allow running the peer application more than
once.