Gathering of local files
Added code that gathers all files in the "file" folder and stores it in
a List, which will be eventually sent to the server when appending a
peer object with an "available files" list in the object. This will all
happen automatically, so if you didn't want some files to be sent as
soon as the app starts, some verification would need to be implemented.