File directory names sent over to server with ease
Problem is, I don't know how to stop it yet and get out of the loop
where it wants more input strings for that specific list of directories.
I thought sending a string like FILESEND would do something to break the
loop but it's probably not that simple.
Anyways, after this, the file names need to be stored in a list
connected to the Peer object. The thread can't act as the Peer object I
don't think, so an entirely new class needs to be implemented, where
multiple instances of the class are held by the server outside of the
individual peer threads. That way, multiple peers should be able to
query one place with all of the peer/file directory information.