Fixed it
Alright. Fixed the null thread issue. I tried to call a system.out on
something that couldn't use .toString(), heh.

So now, it's a matter of a static array holding the Peer constructors.
This information has to be created during the file-appending for loop
that you can see in both the Client and Server classes. After that, the
chat input has been repurposed to send a search query, which still needs
to be properly handled on the Server side.

Neat, huh? It's nice when things actually start to work.