A proper search that works
-You can now search for files known about on the Server(However, this
includes files that can belong to the client, and it has to be typed
explicitly the same string in order to get a result. So readme.txt has
to be typed to completion in order to be found. You also cannot find
multiple results at this time).
-Returns unclickable list panel if the query isn't resolved
For debugging purposes, some extra information about each search query
is included in the Server class log.