Things are getting interesting: Multiple types of requests
Alright, so there are two different types of requests the Client can
make to the server once a connection is established. 1: The search
query, which is primed with a specific string to indicate to the server
that the next bit of information should be searched. Easy peasy.
2: Double click indicating a download query. This primes the server to
know that the next bit of information passed should be downloaded to the
requesting peer. So at this point, the server should figure out who owns
the file, let the owner of the file know "that" is the file the other
peer wishes to download, and then send ip information to both parties to
establish a connection. Easy, right? Right. Sure.