Double click on results returns owner information of file
This should be the setup to allow the a socket to be open on either side
to allow to transfer of the file. Honestly, it's probably better to warn
the foreign peer of the incoming peer request for a file before giving
the requesting peer information to attempt a connection. I'll probably
switch around those actions to avoid the possibility of the foreign peer
not being ready to interpret the request.