Beginning of Peer/Peer File Transfer Communication
-Username is now listed in Client log so you know which is which when
running multiple clients
-Stack overflowed some port automation so we can have more than one
client on the same machine, but it also checks if any other applications
are using the range of ports available, so it's a nice thing to have
either way
-Started the beginnings of the Foreign Peer Listener Loop. Similar to
how the Server works, this will be responsible for recieving peer
requests and throwing them into their own thread for the file transfer
process
-Beginning of a try/catch attempt to connect to the specified peer to
download something has also been started
-The "port" variable stored in the Server class is actually the foreign
port listener for that specific Peer. This is stored because, when a
peer makes a request for a file, they not only need to know the ip, but
the port as well in order to find what the other peer is listening on.

This is incredibly confusing and terrifying. But against all odds, it's
not on fire yet. "Yet".