HERE COMES THE BUGS OH BOY
-There's a syncing issue with information being transferred during the
"request download" double click method. Requires a couple of clicks
before the proper method executes when multiple peers are online.
-Here is the gripe of the current system. Since it's all being hosted on
the same machine, more than one peer can attempt to accept the incoming
connection, "even" if the connecting peer is itself. In other words,
while your attempting to connect, your constantly running socket
searching for a connection can connect itself by accident since it's all
running on the local port.
What needs to happen is the connecting peers post is the same as the
peers current port, it should reject the connection. I don't think this
is too terrible to implement. Hopefully, by doing this, the other
connecting peer should allow the connection to be instantiated.